pyproject.toml
[tool.poetry]
package-mode = false
name = "mereb-challenge"
version = "0.1.0"
description = "Creating a strong auth including social logins using drf and django"
authors = ["degisew <jegisew21@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
django-allauth = "^65.0.2"
jwt = "^1.3.1"
requests = "^2.32.3"
dj-rest-auth = "^6.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
django-debug-toolbar = "^4.4.6"
pytest-django = "^4.9.0"
django-environ = "^0.11.2"
selenium = "^4.25.0"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
